+++ R/get_credentials.R
@@ -19,7 +19,7 @@ get_credentials <- function(base_url, keyring, change_key, dialog, ...) {
   checkmate::assert_logical(change_key, len = 1)
   checkmate::assert_logical(dialog, len = 1)
 
-  rstudio_dialog <- dialog && rstudioapi::isAvailable()
+  rstudio_dialog <- use_rstudio_credential_dialog(dialog)
   test_mode <- getOption("eatPrepTBA.test_mode")
 
   name_prompt <- stringr::str_glue("Enter your username for {base_url}: ")
@@ -36,11 +36,7 @@ get_credentials <- function(base_url, keyring, change_key, dialog, ...) {
       }
 
       if (rstudio_dialog) {
-        name <- rstudioapi::showPrompt(
-          title = "Login",
-          message = name_prompt,
-          default = ""
-        )
+        name <- rstudio_prompt_username(name_prompt)
       } else {
         name <- readline(name_prompt)
       }
@@ -52,12 +48,8 @@ get_credentials <- function(base_url, keyring, change_key, dialog, ...) {
   } else {
     if (is.null(test_mode) || ! test_mode) {
       if (rstudio_dialog) {
-        name <- rstudioapi::showPrompt(
-          title = "Login",
-          message = name_prompt,
-          default = ""
-        )
-        password <- rstudioapi::askForPassword(password_prompt)
+        name <- rstudio_prompt_username(name_prompt)
+        password <- rstudio_prompt_password(password_prompt)
       } else {
         name <- readline(prompt = name_prompt)
         password <- readline(prompt = password_prompt)
@@ -86,4 +78,43 @@ get_credentials <- function(base_url, keyring, change_key, dialog, ...) {
   )
 
   return(credentials)
+}
+
+use_rstudio_credential_dialog <- function(dialog) {
+  if (!dialog) {
+    return(FALSE)
+  }
+
+  is_rstudio_env <- identical(Sys.getenv("RSTUDIO"), "1")
+  is_rstudio_api_available <- tryCatch(
+    {
+      is_available_args <- names(formals(rstudioapi::isAvailable))
+      if ("child_ok" %in% is_available_args) {
+        rstudioapi::isAvailable(child_ok = TRUE)
+      } else {
+        rstudioapi::isAvailable()
+      }
+    },
+    error = function(cnd) FALSE
+  )
+
+  is_rstudio_env || isTRUE(is_rstudio_api_available)
+}
+
+rstudio_prompt_username <- function(prompt) {
+  tryCatch(
+    rstudioapi::showPrompt(
+      title = "Login",
+      message = prompt,
+      default = ""
+    ),
+    error = function(cnd) rstudio_prompt_password(prompt)
+  )
+}
+
+rstudio_prompt_password <- function(prompt) {
+  tryCatch(
+    rstudioapi::askForPassword(prompt),
+    error = function(cnd) readline(prompt = prompt)
+  )
 }